requirements.txt
# Core NLP and ML libraries
pandas==2.0.3
numpy==1.24.3
scipy==1.11.0
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.1
transformers==4.30.2
torch==2.0.1

# Text processing
textblob==0.17.1
wordcloud==1.9.2
langdetect==1.0.9
matplotlib==3.7.2
seaborn==0.12.2

# Data handling
openpyxl==3.1.2
python-docx==0.8.11
PyPDF2==3.0.1
tabulate==0.9.0

# OpenAI-compatible client (UTC platform / LM Studio)
openai>=1.0.0

# LLM Orchestration & RAG
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.2

# Visualization and interface
plotly==5.15.0
streamlit==1.25.0
dash==2.11.1

# Development and testing
jupyter==1.0.0
pytest==7.4.0
black==23.7.0
flake8==6.0.0

# Additional utilities
tqdm==4.65.0
python-dotenv==1.0.0

# PDF extraction (LLM-powered pipeline)
pymupdf>=1.23.0

# FastAPI backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0